+++ services/transactionService.ts
@@ -3,13 +3,15 @@ import { Address, Prisma, Transaction } from '@prisma/client'
 import { syncTransactionsForAddress, subscribeAddresses } from 'services/blockchainService'
 import { fetchAddressBySubstring, fetchAddressById, fetchAddressesByPaybuttonId } from 'services/addressService'
 import { QuoteValues, fetchPricesForNetworkAndTimestamp } from 'services/priceService'
-import { RESPONSE_MESSAGES, USD_QUOTE_ID, CAD_QUOTE_ID, N_OF_QUOTES, KeyValueT, UPSERT_TRANSACTION_PRICES_ON_DB_TIMEOUT } from 'constants/index'
+import { RESPONSE_MESSAGES, USD_QUOTE_ID, CAD_QUOTE_ID, N_OF_QUOTES, KeyValueT, UPSERT_TRANSACTION_PRICES_ON_DB_TIMEOUT, DEFAULT_QUOTE_SLUG, SUPPORTED_QUOTES } from 'constants/index'
 import { productionAddresses } from 'prisma/seeds/addresses'
 import { appendTxsToFile } from 'prisma/seeds/transactions'
 import _ from 'lodash'
 import { CacheSet } from 'redis/index'
 import { SimplifiedTransaction } from 'ws-service/types'
 import { OpReturnData } from 'utils/validators'
+
+type SupportedQuotesType = typeof SUPPORTED_QUOTES[number]
 
 export async function getTransactionValue (transaction: TransactionWithPrices): Promise<QuoteValues> {
   const ret: QuoteValues = {
@@ -488,3 +490,43 @@ export async function fetchTransactionsByPaybuttonId (paybuttonId: string): Prom
 
   return transactions
 }
+
+export const getTransactionValueInCurrency = (transaction: TransactionWithAddressAndPrices, currency?: SupportedQuotesType): number | undefined => {
+  try {
+    const isCurrencyEmptyOrUndefined = currency === '' || currency === undefined
+    const stringCurrency = isCurrencyEmptyOrUndefined ? DEFAULT_QUOTE_SLUG : currency
+    const { prices, amount } = transaction
+
+    const result: { [key in SupportedQuotesType]: number } = {}
+
+    for (const p of prices) {
+      if (p.price.quoteId === USD_QUOTE_ID) {
+        result.usd = p.price.value.times(amount).toNumber()
+      }
+      if (p.price.quoteId === CAD_QUOTE_ID) {
+        result.cad = p.price.value.times(amount).toNumber()
+      }
+    }
+    return result[stringCurrency]
+  } catch (error: any) {
+    console.log(error.message)
+  }
+}
+
+// export const getTransactionValueInCurrency = (transaction: TransactionWithAddressAndPrices, currency?: SupportedQuotesType): number => {
+//   const isCurrencyEmptyOrUndefined = currency === '' || currency === undefined
+//   const stringCurrency: string = isCurrencyEmptyOrUndefined ? '' : DEFAULT_QUOTE_SLUG
+
+//   const { prices, amount } = transaction
+//   const result: { [key in SupportedQuotesType]: number } = {}
+
+//   for (const p of prices) {
+//     if (p.price.quoteId === USD_QUOTE_ID) {
+//       result.usd = p.price.value.times(amount).toNumber()
+//     }
+//     if (p.price.quoteId === CAD_QUOTE_ID) {
+//       result.cad = p.price.value.times(amount).toNumber()
+//     }
+//   }
+//   return (isCurrencyEmptyOrUndefined ? result.usd : result[stringCurrency])
+// }
